+++ crates/cli/src/commands/bin.rs
@@ -1,12 +1,8 @@
 use crate::tools::create_tool;
-use proto_core::{detect_version, VersionType};
+use proto_core::{detect_version, Id, VersionType};
 use starbase::SystemResult;
 
-pub async fn bin(
-    tool_id: String,
-    forced_version: Option<VersionType>,
-    use_shim: bool,
-) -> SystemResult {
+pub async fn bin(tool_id: Id, forced_version: Option<VersionType>, use_shim: bool) -> SystemResult {
     let mut tool = create_tool(&tool_id).await?;
     let version = detect_version(&tool, forced_version).await?;