+++ crates/warpgate/src/lib.rs
@@ -1,11 +1,13 @@
 mod api;
 mod error;
 mod helpers;
+mod id;
 mod loader;
 mod locator;
 mod plugin;
 
 pub use error::*;
+pub use id::*;
 pub use loader::*;
 pub use locator::*;
 pub use plugin::*;