+++ crates/warpgate/tests/loader_test.rs
@@ -1,6 +1,6 @@
 use starbase_sandbox::{create_empty_sandbox, locate_fixture, Sandbox};
 use std::path::PathBuf;
-use warpgate::{GitHubLocator, PluginLoader, PluginLocator};
+use warpgate::{GitHubLocator, Id, PluginLoader, PluginLocator};
 
 fn create_loader() -> (Sandbox, PluginLoader) {
     let sandbox = create_empty_sandbox();
@@ -22,8 +22,8 @@ mod loader {
 
             loader
                 .load_plugin(
-                    "test",
-                    &PluginLocator::SourceFile {
+                    Id::raw("test"),
+                    PluginLocator::SourceFile {
                         file: "".into(),
                         path: PathBuf::from("fake-file"),
                     },
@@ -39,8 +39,8 @@ mod loader {
 
             let path = loader
                 .load_plugin(
-                    "test",
-                    &PluginLocator::SourceFile {
+                    Id::raw("test"),
+                    PluginLocator::SourceFile {
                         file: "".into(),
                         path: fixture.join("test.wasm"),
                     },
@@ -66,7 +66,9 @@ mod loader {
             let (_sandbox, loader) = create_loader();
 
             loader
-                .load_plugin("test", &PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/download/v0.0.2/deno_plugin_invalid_name.wasm".into() })
+                .load_plugin(
+                    Id::raw("test"),
+                    PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/download/v0.0.2/deno_plugin_invalid_name.wasm".into() })
                 .await
                 .unwrap();
         }
@@ -76,7 +78,9 @@ mod loader {
             let (sandbox, loader) = create_loader();
 
             let path = loader
-                .load_plugin("test", &PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/download/v0.0.2/deno_plugin.wasm".into() })
+                .load_plugin(
+                    Id::raw("test"),
+                    PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/download/v0.0.2/deno_plugin.wasm".into() })
                 .await
                 .unwrap();
 
@@ -88,7 +92,9 @@ mod loader {
             let (sandbox, loader) = create_loader();
 
             let path = loader
-                .load_plugin("test", &PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/latest/download/deno_plugin.wasm".into() })
+                .load_plugin(
+                    Id::raw("test"),
+                    PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/latest/download/deno_plugin.wasm".into() })
                 .await
                 .unwrap();
 
@@ -106,8 +112,8 @@ mod loader {
 
             loader
                 .load_plugin(
-                    "test",
-                    &PluginLocator::GitHub(GitHubLocator {
+                    Id::raw("test"),
+                    PluginLocator::GitHub(GitHubLocator {
                         file_prefix: "bun_plugin.wasm".into(),
                         repo_slug: "moonrepo/invalid-repo".into(),
                         tag: None,
@@ -123,8 +129,8 @@ mod loader {
 
             let path = loader
                 .load_plugin(
-                    "test",
-                    &PluginLocator::GitHub(GitHubLocator {
+                    Id::raw("test"),
+                    PluginLocator::GitHub(GitHubLocator {
                         file_prefix: "bun_plugin.wasm".into(),
                         repo_slug: "moonrepo/bun-plugin".into(),
                         tag: Some("v0.0.3".into()),
@@ -142,8 +148,8 @@ mod loader {
 
             let path = loader
                 .load_plugin(
-                    "test",
-                    &PluginLocator::GitHub(GitHubLocator {
+                    Id::raw("test"),
+                    PluginLocator::GitHub(GitHubLocator {
                         file_prefix: "bun_plugin.wasm".into(),
                         repo_slug: "moonrepo/bun-plugin".into(),
                         tag: None,